Extract createApplication
